requirements.txt
google-generativeai
PyQt5
pyttsx3
sounddevice
scipy